Moved postMessageHtmlToComponent post_message_view to post_utils for use in textbox
